Moved some files and implemented naming conventions
