+++ next/src/app/layout.jsx
@@ -12,8 +12,8 @@ import '/src/styles/checkout.css';
 import '/src/styles/complete.css';
 import '/src/styles/product.css';
 
-import SentryQueryInitializer from '../../ui/sentry-query-initializer';
-import HomeContent from '../../ui/home-content';
+import SentryQueryInitializer from '../ui/SentryQueryInitializer';
+import HomeContent from '../ui/HomeContent';
 
 export default function RootLayout({
   // Layouts must accept a children prop.
