+++ next/src/ui/SentryQueryInitializer.jsx
@@ -3,7 +3,6 @@
 'use client';
 
 import React from 'react';
-import { crasher } from '/src/utils/errors';
 
 import { useSearchParams } from 'next/navigation';
 import * as Sentry from '@sentry/nextjs';
